Check version against published version
- Check version against published version
- Turn on version check before deploying master build
- Bump version to v0.28.7
