+++ src/implementations/react/scripts/validate-version/getLocalVersion.js
@@ -1,0 +1,6 @@
+const localPackageJson = require("../../package.json");
+
+// Get the package.json version from the local codebase
+module.exports = function getLocalVersion() {
+  return localPackageJson.version;
+};
